整合两个人的任务

+++ src/main/java/com/serching/fulltextsearching/entity/TOperationLog.java
@@ -8,10 +8,10 @@ import java.time.LocalDateTime;
 @TableName("t_operation_log")
 public class TOperationLog {
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     private String title; // 操作描述
     private Integer operationType; // 操作类型
-    private Integer objectId; // 操作对象id:如知识库id、知识文档id
+    private Long objectId; // 操作对象id:如知识库id、知识文档id
     private String objectType; // 知识库、知识、成员
     private Long createdBy; // 创建人
     private LocalDateTime createdAt;
